requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
xlrd
msoffcrypto-tool
python-dateutil
